Initial support for GCC 4.5 compatibility.
Not-yet-implemented keywords are not handled in the header (just in the
tests), you have to provide something similar to the following and be
sure that it won't collide with anything anywhere else:

    #define constexpr
    #define noexcept
    #define nullptr 0

Because of this, the 4.4/4.5 compatibility isn't something to be taken
very seriously, just an fugly workaround to get this to compile on old
systems. It is not meant to enable actual development on them, as
some things working with these compilers may not work on newer systems.
But things working on newer compilers should be working on the older
ones, which is the actual point of this compatibility branch.